88 protected routes middleware

+++ src/app/api/events/[eventID]/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import connectDB from "@database/db";
 import EventSchema from "@database/eventSchema";
+import {getSession} from "services/auth/cookietoUsertype"
 
 type IParams = {
   params: {
@@ -24,8 +25,14 @@ export async function GET(req: NextRequest, { params }: IParams) {
 
 export async function DELETE(req: NextRequest, { params }: IParams) {
   await connectDB();
-  const { eventID } = params;
+  // makes route exclusive to admin
+  let usertype;
+  usertype = await getSession(req);
+  if (usertype != 'admin'){
+    return NextResponse.json(`Forbidden Action`, {  status: 400, });
+  }
 
+  const { eventID } = params;
   try {
     await EventSchema.deleteOne({ eventID }).orFail();
     return NextResponse.json(`Event deleted.`, { status: 200 });
